requirements_test.txt
# Warning: For automatic run of test in Gitlab CI, we must not include other things that pytest-homeassistant-custom-component
# full reload dependencies: pip cache purge && pip install --use-feature=fast-deps -r requirements_test.txt
# list all version of pytest-homeassistant-custom-component: pip index versions pytest-homeassistant-custom-component
-r requirements_dev.txt

# Core Testing Framework
pytest
pytest-asyncio
pytest-homeassistant-custom-component
coverage

# DNS Fixes (Keep only if you still encounter specific CI failures)
aiodns==3.5.0
pycares==4.11.0
